Resize clothing textures with a dropdown or when the overlay is larger than the original texture (#35)
Just like with the body over/underlays, resize the texture of the clothing when the settings is enabled and the applied overlay/color mask is larger than the original.
This is also done when the color mask is larger because of patterns.

Also fixed a small problem where patterns would disappear when changing color mask until you edited a pattern setting

* Re-apply patterns after changing color mask

* Resize clothing textures

* Add dropdowns to override texture size

* Add Save/load

* Change texture size immediatly when setting texture

* Only update the patterns that are active, if any

* Only update patterns when needed

* Change some formatting

* Clean stored data and try to keep it clean