Optimize loading overlays in studio (#36)
Seems to work fine like this instead of forcing the character to load the current outfit again, despite the warning about the dynamic bones messing up.

I tested it myself on both KK and KKS, and a few others also tested it for me on either KK or KKS and it didn't break for anybody.

The aggressive refresh is still performed in studio when loading any type of mask.