[AI/HS2] Add DetailMainTex overlay support for face and body skin (#37)
This feature adds support for overlays and underlays that apply specifically to the skin DetailMainTex texture - this is the metallic/gloss map for the skin (face and body). 4 new texture slots are added to the skin overlays menu:

face detail overlay
body detail overlay
face detail underlay
body detail underlay
Similarly to the main skin texture, the DetailMainTex texture is generated dynamically by the game, as it combines the corresponding metallic/gloss maps for things like tattoos, lips, eye shadow and blushes into one texture. Thus, compared to editing DetailMainTex in the Material Editor, this approach allows the user to blend custom maps with the game-generated texture, instead of always completely replacing it. The underlays get applied before the tattoos, lips, eye shadow and blushes are applied, while overlays apply on top of those. The alpha channel acts as a mask.

I wasn't able to find an equivalent in KK: in AI/HS2, DetailMainTex is always stored as the 2nd element in a CustomTextureCreate array (with albedo being the 1st), but for the other games there seems to be only one CustomTextureCreate (the albedo). So I limited the functionality and TexType values to AI/HS2.