Pattern support (#38)
Adds the ability to override the pattern with a custom one. It only works when a custom injected pattern is selected, which uses a place holder pattern telling to user to replace it.

I also fixed the copying of clothes not copying colormask overrides (and thus also now pattern overrides)

* Get pattern override working
* Add patterns to tops
* Dump original pattern
* Original pattern dumping for HS2/AI
* Clean dumping code a little
* Fix copy/paste oversight
* Add custom overlay pattern you need to select
* HS2/AI compatibility
* Fix clothes copying
* Also overwrite thumbnail without MakerOptimizations
* Fix null ref exception in hook for HS2
* Make patch optional
* Keep old method as obsolete overload