Add new overlay blending mode for clothes (#41)
The old alpha blending done though the shader was wrong and thus resulted in wrong alpha values. So if you were to dump the original texture of a pantyhose and re-import it as an override texture, it would be more transparent and darker.
The new blending fixes this.

The new blending is only used for newly added overlays and old overlays will keep using the old blending method for backwards compatibility.
The old blending mode is also still used for the skin/eye over/underlays. From my testing it seems this blending was only a problem if the source texture also had alpha. And since the new blending doesn't play nice in some cases (KK(S) eye overlays, HS2 body overlays) and the issue didn't occur there anyway, I opted to keep using the old blending there.

Also bumped the PluginData version because of the added BlendingMode enum